rename Event file component to events to not conflict with show page
