+++ src/components/Events.js
@@ -1,7 +1,7 @@
 import React, { useEffect } from 'react';
 import { useDispatch, useSelector } from 'react-redux';
 import { fetchEvents, selectEvents } from '../features/events/eventSlice';
-import styles from './styles/event.module.css';
+import styles from './styles/events.module.css';
 
 export default function Event() {
   const dispatch = useDispatch();
