--- src/components/styles/event.css
@@ -1,18 +1,0 @@
-
-img {
-  width: 300px;
-  height: 400px;
-}
-.events-ul {
- display: grid;
- gap: var(--size-3);
- grid-auto-flow: column;
- grid-auto-columns: 21%;
- overflow-x: scroll;
- overscroll-behavior-inline: contain;
-}
-.events-ul > li {
-  display: flex;
-  flex-direction: column;
-  padding: 10px;
-}
